+++ extensions/authclient/widgets/AuthChoiceStyleAsset.php
@@ -12,19 +12,13 @@ use yii\web\AssetBundle;
 /**
  * AuthChoiceAsset is an asset bundle for [[AuthChoice]] widget.
  *
- * @see AuthChoiceStyleAsset
- *
  * @author Paul Klimov <klimov.paul@gmail.com>
  * @since 2.0
  */
-class AuthChoiceAsset extends AssetBundle
+class AuthChoiceStyleAsset extends AssetBundle
 {
     public $sourcePath = '@yii/authclient/assets';
-    public $js = [
-        'authchoice.js',
-    ];
-    public $depends = [
-        'yii\authclient\widgets\AuthChoiceStyleAsset',
-        'yii\web\YiiAsset',
+    public $css = [
+        'authchoice.css',
     ];
 }